Checking for sim log dir existence

+++ pippin/snana_sim.py
@@ -229,24 +229,28 @@ class SNANASimulation(ConfigBasedExecutable):
                 return Task.FINISHED_FAILURE
         else:
             self.logger.warn(f"Simulation {self.name} logging file does not exist: {self.logging_file}")
-        for file in os.listdir(self.sim_log_dir):
-            if not file.startswith("TMP") or not file.endswith(".LOG"):
-                continue
-            try:
-                with open(self.sim_log_dir + "/" + file, "r") as f:
-                    for line in f.read().splitlines():
-                        if (" ABORT " in line or "FATAL[" in line) and not output_error:
-                            output_error = True
-                            self.logger.error(f"Fatal error in simulation. See {self.sim_log_dir}/{file} for details.")
-                        if output_error:
-                            self.logger.info(f"Excerpt: {line}")
-                if output_error:
-                    self.logger.debug("Removing hash on failure")
-                    os.remove(self.hash_file)
-                    chown_dir(self.output_dir)
-                    return Task.FINISHED_FAILURE
-            except FileNotFoundError:
-                self.logger.warning("Cannot find log file in process of reading it in (maybe it just got zipped up): {self.sim_log_dir + '/' + file}")
+        if os.path.exists(self.sim_log_dir):
+            for file in os.listdir(self.sim_log_dir):
+                if not file.startswith("TMP") or not file.endswith(".LOG"):
+                    continue
+                try:
+                    with open(self.sim_log_dir + "/" + file, "r") as f:
+                        for line in f.read().splitlines():
+                            if (" ABORT " in line or "FATAL[" in line) and not output_error:
+                                output_error = True
+                                self.logger.error(f"Fatal error in simulation. See {self.sim_log_dir}/{file} for details.")
+                            if output_error:
+                                self.logger.info(f"Excerpt: {line}")
+                    if output_error:
+                        self.logger.debug("Removing hash on failure")
+                        os.remove(self.hash_file)
+                        chown_dir(self.output_dir)
+                        return Task.FINISHED_FAILURE
+                except FileNotFoundError:
+                    self.logger.warning("Cannot find log file in process of reading it in (maybe it just got zipped up): {self.sim_log_dir + '/' + file}")
+        else:
+            self.logger.error(f"Sim log dir doesn't exist at {self.sim_log_dir}. Warning, there might be orphaned jobs from this.")
+            return Task.FINISHED_FAILURE
 
         # Check to see if the done file exists
